Create Odd_even_pairs
write a python code such that splits into n/2 pairs, such that sum of integers in each pair is odd. n is even and every element must be present in exactly 1 pair

You can solve this problem by trying to form pairs such that the sum of each pair is odd.

🔸 Observation:

An odd sum can only be obtained by adding an even number and an odd number.

So, we need to pair each even number with an odd number.

✅ Steps:
Separate the numbers into odd and even lists.

If the number of odds ≠ number of evens, such pairing is not possible.

Otherwise, pair each odd with an even.